feat(check): route run/test/cache type-checking through native tsc
`deno check` already runs the pinned native TypeScript compiler (tsgo),
but `deno run --check`, `deno test`, and `deno cache --check` still used
the in-binary forked-TS snapshot via TypeChecker::check -> tsc::exec.
This routes those three onto the native pipeline too - they all converge
on ModuleLoadPreparer::prepare_module_load, so a single call site flips
all of them.

The native check core is extracted so both `deno check` and the
run/test/cache path share it: run_native_check_over_roots in
cli/tsc/native.rs owns the tsconfig build + tsc spawn + diagnostic
remap, and TypeChecker::check_native owns the graph walk, incremental
cache decision, sync-types materialization, and hash record. The
compiler-download deps are bundled into NativeTscInstallDeps and stored
on TypeChecker (built in the factory).

The old path was diagnostics-only for these commands (no tsc emit, no
.tsbuildinfo dependency; Deno emits JS via swc), so native (--noEmit) is
a faithful replacement. The check future is boxed at the call site to
stay under clippy's large_futures threshold on the module-load hot path.

compile --check and publish still use the old path (separate call site,
follow-up); the forked-TS snapshot stays in the binary for the LSP.
